MNT/BUG: Fix callbacks for selections and begin organizing widget tools
