+++ online_model/app/__init__.py
@@ -37,100 +37,6 @@ if PROTOCOL == "pva":
     CONTEXT = Context("pva")
 
 
-def set_pv_from_slider(attr, old, new, pvname: str, scale) -> None:
-    """
-    Callback function for slider change.
-
-    Parameters
-    ----------
-    attr
-
-    old
-
-    new
-
-    pvname: str
-        Process variable name
-
-    scale:
-
-    """
-    if PROTOCOL == "pva":
-        CONTEXT.put(pvname, new * scale)
-
-    elif PROTOCOL == "ca":
-        caput(pvname, new * scale)
-
-
-def build_slider(title: str, pvname, scale, start, end, step) -> Slider:
-    """
-    Utility function for building a slider.
-
-    Parameters
-    ----------
-    title:str
-        Slider title
-
-    pvname:str
-        Process variable name
-
-    scale:float/int
-        Scale of the slider
-
-    start:float
-        Lower range of the slider
-
-    end:float
-        Upper range of the slider
-
-    step:np.float64
-        The step between consecutive values
-
-    Returns
-    -------
-    bokeh.models.widgets.sliders.Slider
-
-    """
-
-    # initialize value
-    start_val = None
-    if PROTOCOL == "pva":
-        start_val = CONTEXT.get(pvname)
-    elif PROTOCOL == "ca":
-        start_val = caget(pvname)
-
-    slider = Slider(
-        title=title, value=scale * start_val, start=start, end=end, step=step
-    )
-
-    slider.on_change("value", partial(set_pv_from_slider, pvname=pvname, scale=scale))
-
-    return slider
-
-
-def build_sliders(CMD_PVDB):
-
-    sliders = []
-
-    for ii, pv in enumerate(CMD_PVDB):
-        title = pv + " (" + CMD_PVDB[pv]["units"] + ")"
-        pvname = PREFIX + ":" + pv
-        step = (CMD_PVDB[pv]["range"][1] - CMD_PVDB[pv]["range"][0]) / 100.0
-        scale = 1
-
-        slider = build_slider(
-            title,
-            pvname,
-            scale,
-            CMD_PVDB[pv]["range"][0],
-            CMD_PVDB[pv]["range"][1],
-            step,
-        )
-        sliders.append(slider)
-
-    return sliders
-
-
 class PVImageMonitor:
     """
     Monitor for updating and formatting image data.
